requirements.txt
requests
click
pillow
pyppeteer